Update Units package to support UTF-8 symbols
UTF-8 symbols for SI multipliers and special symbols are enabled (old Latin-1 definitions are preserved for compatibility).
For "micro", several variants are supported - Unicode mu as Greek letter and as micro sign, conventional prefixes "mc" and "u".

Definition of units in Units.dat is updated to use UTF-8 symbols instead of Latin-1.

Definition of symbol for micron is corrected to be "µm" instead of just "µ".
Definition of oersted and gilbert units is corrected to divide by pi instead of multiplying.

Added test demo draw units